Mask attribute Fixes #3
